DataQuick/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.5.0

# Database (SQLite is built-in, no driver needed)
SQLAlchemy>=2.0.23
alembic>=1.13.0

# Data processing & profiling (use pre-built wheels on Windows)
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0
openpyxl>=3.0.0

# Visualization & lineage
plotly>=5.0.0
matplotlib>=3.7.0
networkx>=3.1
altair>=5.0.0

# LLM & Agent orchestration (optional, with fallback)
langchain>=0.1.0
requests>=2.31.0

# Embeddings & Vector stores
sentence-transformers>=2.2.0
chromadb>=0.4.0
pyvis>=0.3.0

# UI
streamlit>=1.28.0

# Scheduling
APScheduler>=3.10.0

# Utilities
click>=8.1.0
pyyaml>=6.0
loguru>=0.7.0

# Dev & testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
